Separated in different files the encoder and decoder of Huffman Coding
